+++ utils/discord/argument.py
@@ -1,3 +1,5 @@
+from typing import Union
+
 from discord import app_commands
 from discord.app_commands.transformers import ALLOWED_DEFAULTS
 
@@ -6,6 +8,7 @@ ALLOWED_DEFAULTS = set(_type for tup in ALLOWED_DEFAULTS.values() for _type in t
 
 
 class DiscordArgument:
+    # __origin__ = Union
     choices: dict | None = None
 
     @classmethod
